Update and rename main.cpp
